# ip match: simplify and speed up the LC Trie implementation