--- a/source/common/network/lc_trie.cc
+++ b/source/common/network/lc_trie.cc
@@ -14,21 +14,27 @@ LcTrie::LcTrie(const std::vector<std::pair<std::string, std::vector<Address::Cid
   for (const auto& pair_data : tag_data) {
     for (const auto& cidr_range : pair_data.second) {
       if (cidr_range.ip()->version() == Address::IpVersion::v4) {
-        IpPrefix<Ipv4> ip_prefix;
-        ip_prefix.tag_ = pair_data.first;
-        ip_prefix.length_ = cidr_range.length();
-        ip_prefix.ip_ = ntohl(cidr_range.ip()->ipv4()->address());
+        IpPrefix<Ipv4> ip_prefix(ntohl(cidr_range.ip()->ipv4()->address()), cidr_range.length(),
+                                 pair_data.first);
         ipv4_prefixes.push_back(ip_prefix);
       } else {
-        IpPrefix<Ipv6> ip_prefix;
-        ip_prefix.tag_ = pair_data.first;
-        ip_prefix.length_ = cidr_range.length();
-        ip_prefix.ip_ = Utility::Ip6ntohl(cidr_range.ip()->ipv6()->address());
+        IpPrefix<Ipv6> ip_prefix(Utility::Ip6ntohl(cidr_range.ip()->ipv6()->address()),
+                                 cidr_range.length(), pair_data.first);
         ipv6_prefixes.push_back(ip_prefix);
       }
     }
   }
+  BinaryTrie<Ipv4> ipv4_temp;
+  for (const auto& prefix : ipv4_prefixes) {
+    ipv4_temp.insert(prefix);
+  }
+  ipv4_prefixes = ipv4_temp.push_leaves();
   ipv4_trie_.reset(new LcTrieInternal<Ipv4>(ipv4_prefixes, fill_factor, root_branching_factor));
+  BinaryTrie<Ipv6> ipv6_temp;
+  for (const auto& prefix : ipv6_prefixes) {
+    ipv6_temp.insert(prefix);
+  }
+  ipv6_prefixes = ipv6_temp.push_leaves();
   ipv6_trie_.reset(new LcTrieInternal<Ipv6>(ipv6_prefixes, fill_factor, root_branching_factor));
 }