--- a/test/common/network/BUILD
+++ b/test/common/network/BUILD
@@ -2,6 +2,7 @@ licenses(["notice"])  # Apache 2
 
 load(
     "//bazel:envoy_build_system.bzl",
+    "envoy_cc_binary",
     "envoy_cc_test",
     "envoy_cc_test_library",
     "envoy_package",
@@ -182,5 +183,18 @@ envoy_cc_test(
         "//source/common/network:address_lib",
         "//source/common/network:utility_lib",
         "//test/test_common:environment_lib",
+    ],
+)
+
+envoy_cc_binary(
+    name = "lc_trie_speed_test",
+    testonly = 1,
+    srcs = ["lc_trie_speed_test.cc"],
+    external_deps = [
+        "benchmark",
+    ],
+    deps = [
+        "//source/common/network:lc_trie_lib",
+        "//source/common/network:utility_lib",
     ],
 )