*Description*:

This change removes the special-case handling for nested lists.
It works by preprocessing the input list of prefixes into an
equivalent list where no prefix is nested under another.

*Benchmark test results*:

Before:
```
-------------------------------------------------------------------------
Benchmark                                  Time           CPU Iterations
-------------------------------------------------------------------------
BM_LcTrieLookup                          364 ns        364 ns    1916720
BM_LcTrieLookupWithNestedPrefixes      20546 ns      20545 ns      33952
```

After:
```
-------------------------------------------------------------------------
Benchmark                                  Time           CPU Iterations
-------------------------------------------------------------------------
BM_LcTrieLookup                          233 ns        233 ns    3008092
BM_LcTrieLookupWithNestedPrefixes        411 ns        411 ns    1702732
```

*Risk Level*: Medium
*Testing*: unit tests
*Docs Changes*: N/A
*Release Notes*: N/A
Fixes #3188

Signed-off-by: Brian Pane <bpane@pinterest.com>