@brian-pane 2018-05-02
The error in the CircleCI Mac test run is in a test called `IpVersions/SslSocketTest.HalfClose/IPv6`, and this PR doesn't change any socket code, so I think that error is unrelated.