@htuch 2018-05-03
@brian-pane will get to this first thing tomorrow, sorry about the delay, want to make sure I have cycles to fully understand this.