@ggreenway 2018-05-04
@brian-pane Just saw this test failure in a build a few after this was merged. Can you please take a look and see if it's related to this change?

https://circleci.com/gh/envoyproxy/envoy/49127?utm_campaign=vcs-integration-link&utm_medium=referral&utm_source=github-build-link
```
FAIL: @envoy//test/common/network:lc_trie_test (see /build/tmp/_bazel_bazel/400406edc57d332f0b9b805d2b8e33a1/execroot/envoy/bazel-out/k8-dbg/testlogs/external/envoy/test/common/network/lc_trie_test/test.log)

INFO: From Testing @envoy//test/common/network:lc_trie_test:

==================== Test output for @envoy//test/common/network:lc_trie_test:
[==========] Running 14 tests from 1 test case.
[----------] Global test environment set-up.
[----------] 14 tests from LcTrieTest
[ RUN      ] LcTrieTest.IPv4Defaults
[       OK ] LcTrieTest.IPv4Defaults (301 ms)
[ RUN      ] LcTrieTest.RootBranchingFactor
[       OK ] LcTrieTest.RootBranchingFactor (367 ms)
[ RUN      ] LcTrieTest.IPv4AddressSizeBoundaries
[       OK ] LcTrieTest.IPv4AddressSizeBoundaries (389 ms)
[ RUN      ] LcTrieTest.IPv4Boundaries
[       OK ] LcTrieTest.IPv4Boundaries (639 ms)
[ RUN      ] LcTrieTest.IPv6
[       OK ] LcTrieTest.IPv6 (399 ms)
[ RUN      ] LcTrieTest.IPv6AddressSizeBoundaries
[       OK ] LcTrieTest.IPv6AddressSizeBoundaries (364 ms)
[ RUN      ] LcTrieTest.IPv6Boundaries
[       OK ] LcTrieTest.IPv6Boundaries (602 ms)
[ RUN      ] LcTrieTest.CatchAllIPv4Prefix
[       OK ] LcTrieTest.CatchAllIPv4Prefix (658 ms)
[ RUN      ] LcTrieTest.CatchAllIPv6Prefix
[       OK ] LcTrieTest.CatchAllIPv6Prefix (708 ms)
[ RUN      ] LcTrieTest.BothIpvVersions
[       OK ] LcTrieTest.BothIpvVersions (780 ms)
[ RUN      ] LcTrieTest.NestedPrefixes
[       OK ] LcTrieTest.NestedPrefixes (677 ms)
[ RUN      ] LcTrieTest.NestedPrefixesWithCatchAll
[       OK ] LcTrieTest.NestedPrefixesWithCatchAll (671 ms)
[ RUN      ] LcTrieTest.MaximumEntriesException
[       OK ] LcTrieTest.MaximumEntriesException (47299 ms)
[ RUN      ] LcTrieTest.ExceedingAllocatedTrieNodesException
Killed
================================================================================```