@brian-pane 2018-05-04
@ggreenway thanks for pointing that out I'm checking now, and I'll post an update shortly.