@brian-pane 2018-05-04
@ggreenway I can't repro the error, but I think I know why it happens:

With this PR, the intermediate binary trie is built _before_ the check for too many trie nodes. In a way, that makes sense: the LC trie has a small address space because it only uses 20-bit "pointers," so it needs to worry about running out of address space; but the binary trie uses native pointers and won't run out of address space.

However, on server or container with a small amount of memory, this test (which tries to build an enormous LC trie) can run out of memory while building the intermediate binary trie.

So I'll make a follow-up PR that moves the too-many-prefixes check before the creation of the intermediate binary trie.